ref(sendInvitesForItems): convert to action.
